Alert

+++ components/FavoriteComponent.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import {View,FlatList} from 'react-native';
+import {View,FlatList,Alert} from 'react-native';
 import {Tile,ListItem} from 'react-native-elements';
 import { connect } from 'react-redux';
 import { baseUrl } from '../shared/baseUrl';
@@ -25,7 +25,24 @@ class Favorites extends Component{
                 {
                     text:"Delete",
                     type: 'delete',
-                    onPress:()=>this.props.deleteFavorite(item.id)
+                    onPress:()=>{
+                        Alert.alert(
+                            'Delete Favorite',
+                            'Are you sure you wish to delete the favorite dish'+item.name+' ?',
+                            [
+                                {
+                                    text:'Cancel' ,
+                                    onPress:()=>console.log(item.name+'Not deleted'),
+                                    style: 'cancel'
+                                },
+                                {
+                                    text:'Ok',
+                                    onPress:()=>this.props.deleteFavorite(item.id),
+                                }
+                            ],
+                            {cancelable:false}
+                        )
+                    }
                 }
             ];
             return(
